moved quadtree to enhancements
