+++ liquidfun/Contributions/Enhancements/Controllers/b2Controller.h
@@ -19,8 +19,8 @@
 #ifndef B2_CONTROLLER_H
 #define B2_CONTROLLER_H
 
-#include "../../Dynamics/b2World.h"
-#include "../../Dynamics/b2Body.h"
+#include "../../../Box2D/Box2D/Dynamics/b2World.h"
+#include "../../../Box2D/Box2D/Dynamics/b2Body.h"
 
 class b2Body;
 class b2World;
